+++ src/lua/mod.rs
@@ -16,6 +16,7 @@ mod mud;
 mod regex;
 mod script;
 mod store_data;
+mod timer;
 mod tts;
 mod ui_event;
 pub mod util;